requirements.txt
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
distro==1.5.0
gpiozero==1.6.2
idna==2.10
numpy==1.23.0
pandas==2.1.4
pidng==4.0.9
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
pytz==2023.3.post1
requests==2.25.1
rpi-lcd==0.0.3
RPi.GPIO==0.7.0
six==1.16.0
smbus==1.1.post2
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
tzdata==2023.3
urllib3==1.26.5
v4l2-python3==0.3.2
